requirements.txt
pytest == 8.3.4
coverage == 7.6.9

sqlalchemy == 2.0.35
pydantic~=2.10.4

passlib==1.7.3
bcrypt==3.2.0
PyJWT ~= 2.10.1

# FastAPI is a modern, fast (high-performance), web framework for building APIs with Python 3.6+ based on standard Python type hints.
starlette~=0.41.3
fastapi~=0.115.6
uvicorn~=0.34.0
python-multipart ~=0.0.20

